split swearjar node from browser
