+++ package.json
@@ -28,6 +28,7 @@
     }
   ],
   "main": "lib/swearjar-node",
+  "browser": "lib/swearjar-browser",
   "engines": {
     "node": ">= 0.8.0"
   },
